[feat] Add drag'n drop to tabs and icons to buttons
Upgraded the javascript funtion to show tabs, adding the possibility to
drag the tabs with the mouse, rearanging them. It was added bootstrap
icons to the buttons, leading to a more sleek user interface. For
stilystic purposes, a separator was added to the space between the
url/search bar and the tab list.