[feat] Remove search bar and add rounded corners
Removed the search bar in favour of the browser's native url bar. This
makes so that no further problems related to the url bar occur. It was
also added inverse rounded corners to the top and bottom right, to make
it look more integrated with the rest of the browser.